reformat cache directory
