+++ contracts/programs/data-feeds-cache/src/instructions/mod.rs
@@ -1,0 +1,29 @@
+pub mod accept_ownership;
+pub mod close_decimal_report;
+pub mod init_decimal_reports;
+pub mod init_legacy_feeds_config;
+pub mod initialize;
+pub mod on_report;
+pub mod preview_decimal_feed_configs;
+pub mod query_feed_metadata;
+pub mod query_values;
+pub mod set_decimal_feed_configs;
+pub mod set_feed_admin;
+pub mod transfer_ownership;
+pub mod update_forwarder_id;
+pub mod update_legacy_feeds_config;
+
+pub use accept_ownership::*;
+pub use close_decimal_report::*;
+pub use init_decimal_reports::*;
+pub use init_legacy_feeds_config::*;
+pub use initialize::*;
+pub use on_report::*;
+pub use preview_decimal_feed_configs::*;
+pub use query_feed_metadata::*;
+pub use query_values::*;
+pub use set_decimal_feed_configs::*;
+pub use set_feed_admin::*;
+pub use transfer_ownership::*;
+pub use update_forwarder_id::*;
+pub use update_legacy_feeds_config::*;
