+++ contracts/programs/data-feeds-cache/src/lib.rs
@@ -1,3 +1,8 @@
+pub mod instructions;
+pub mod utils;
+
+pub use instructions::*;
+
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program::program::invoke_signed;
 use std::cell::RefMut;
@@ -12,17 +17,13 @@ mod event;
 pub mod state;
 
 use anchor_lang::solana_program::{hash, system_instruction};
-use common::{ANCHOR_DISCRIMINATOR, MAX_WORKFLOW_METADATAS, ZERO_ADDRESS, ZERO_DATA_ID};
+use common::{MAX_WORKFLOW_METADATAS, ZERO_ADDRESS, ZERO_DATA_ID};
 use context::*;
 use error::{AuthError, DataCacheError};
-use event::{DecimalFeedConfigSet, LegacyFeedsReported};
 use state::{
     AccountList, CacheTransmission, DecimalReport, FeedConfig, LegacyFeedEntry, LegacyFeedsConfig,
-    ReceivedDecimalReport, WorkflowMetadata, WritePermissionFlag,
+    WorkflowMetadata,
 };
-
-// derived from hash::hash("global:cache_submit".as_bytes()).to_bytes()[..8].to_vec()
-pub const SUBMIT_DISCRIMINATOR: [u8; 8] = [173, 69, 171, 96, 179, 143, 243, 226];
 
 /// Data Feed Cache receives report updates from an pre-authorized Forwarder authority (sender)
 /// only for pre-authorized data ids and workflows.
@@ -30,111 +31,25 @@ pub const SUBMIT_DISCRIMINATOR: [u8; 8] = [173, 69, 171, 96, 179, 143, 243, 226]
 /// and the authorization of workflows.
 #[program]
 pub mod data_feeds_cache {
-
-    use anchor_lang::{
-        prelude::borsh::BorshSerialize,
-        solana_program::{
-            instruction::Instruction,
-            program::{invoke, invoke_signed},
-            system_instruction,
-        },
-        Discriminator,
-    };
-
-    use crate::event::{
-        CacheInitialized, DecimalReportClosed, DecimalReportInitialized, DecimalReportUpdated,
-        FeedAdminUpdated, ForwarderUpdated, InvalidUpdatePermission, LegacyFeedsConfigInitialized,
-        LegacyFeedsConfigUpdated, OwnershipAcceptance, OwnershipTransfer, StaleDecimalReport,
-    };
-
     use super::*;
 
     /// Creates a new data cache instance with a dedicated state account.
     /// Sets the initial feed admins and state configuration.
     pub fn initialize(ctx: Context<Initialize>, feed_admins: Vec<Pubkey>) -> Result<()> {
-        let state = &mut ctx.accounts.state.load_init()?;
-        state.owner = ctx.accounts.owner.key();
-        state.forwarder_id = ctx.accounts.forwarder_program.key();
-
-        let mut prev_admin = Pubkey::default();
-        for admin in feed_admins.iter() {
-            require!(
-                &prev_admin < admin,
-                DataCacheError::AddressesMustStrictlyIncrease
-            );
-            state.feed_admins.push(*admin);
-            emit!(FeedAdminUpdated {
-                state: ctx.accounts.state.key(),
-                admin: admin.clone(),
-                is_admin: true,
-            });
-            prev_admin = *admin;
-        }
-
-        let (_, bump) = Pubkey::find_program_address(
-            &[b"legacy_writer", ctx.accounts.state.key().as_ref()],
-            &crate::ID,
-        );
-        state.legacy_writer_bump = bump;
-
-        emit!({
-            CacheInitialized {
-                state: ctx.accounts.state.key(),
-                forwarder_id: state.forwarder_id,
-                legacy_writer_bump: state.legacy_writer_bump,
-            }
-        });
-
-        Ok(())
+        initialize::handler(ctx, feed_admins)
     }
 
     /// Not to be used in normal circumstances unless the original forwarder
     /// undergoes maintenance.
     pub fn update_forwarder_id(ctx: Context<UpdateForwarder>) -> Result<()> {
-        let mut state = ctx.accounts.state.load_mut()?;
-
-        emit!({
-            ForwarderUpdated {
-                previous_forwarder: state.forwarder_id,
-                new_forwarder: ctx.accounts.forwarder_program.key(),
-            }
-        });
-
-        state.forwarder_id = ctx.accounts.forwarder_program.key();
-
-        Ok(())
+        update_forwarder_id::handler(ctx)
     }
 
     /// Add or remove a feed admin.
     /// Feed Admins are highly priveleged roles who can unilaterally edit the configuration of feeds
     /// and the authorization of workflows.
     pub fn set_feed_admin(ctx: Context<SetFeedAdmin>, admin: Pubkey, is_admin: bool) -> Result<()> {
-        require_keys_neq!(admin, Pubkey::default(), DataCacheError::InvalidAddress);
-
-        let mut state = ctx.accounts.state.load_mut()?;
-        let mut changed = false;
-
-        match (is_admin, state.feed_admins.binary_search(&admin)) {
-            (false, Ok(i)) => {
-                state.feed_admins.remove(i);
-                changed = true;
-            },
-            (true, Err(i)) => {
-                state.feed_admins.insert(i, admin);
-                changed = true;
-            },
-            _ => {},
-        }
-
-        if changed {
-            emit!(FeedAdminUpdated {
-                state: ctx.accounts.state.key(),
-                admin,
-                is_admin
-            });
-        }
-
-        Ok(())
+        set_feed_admin::handler(ctx, admin, is_admin)
     }
 
     /// Step 1 of 2-step ownership process: propose a new owner
@@ -142,37 +57,12 @@ pub mod data_feeds_cache {
         ctx: Context<TransferOwnership>,
         proposed_owner: Pubkey,
     ) -> Result<()> {
-        let state = &mut ctx.accounts.state.load_mut()?;
-        require!(
-            proposed_owner != Pubkey::default()
-                && proposed_owner != state.owner
-                && proposed_owner != state.proposed_owner,
-            DataCacheError::InvalidProposedOwner
-        );
-
-        state.proposed_owner = proposed_owner;
-        emit!(OwnershipTransfer {
-            state: ctx.accounts.state.key(),
-            current_owner: state.owner,
-            proposed_owner
-        });
-
-        Ok(())
+        transfer_ownership::handler(ctx, proposed_owner)
     }
 
     /// Step 2 of 2-step ownership process: accept ownership
     pub fn accept_ownership(ctx: Context<AcceptOwnership>) -> Result<()> {
-        let state = &mut ctx.accounts.state.load_mut()?;
-
-        emit!(OwnershipAcceptance {
-            state: ctx.accounts.state.key(),
-            previous_owner: state.owner,
-            new_owner: state.proposed_owner
-        });
-
-        state.owner = std::mem::take(&mut state.proposed_owner);
-
-        Ok(())
+        accept_ownership::handler(ctx)
     }
 
     /// Closes the data id's associated decimal report account and feed config account.
@@ -180,24 +70,7 @@ pub mod data_feeds_cache {
     /// `set_decimal_feeds_config` prior to clear out the workflow metadata and
     /// close write permission flag accounts
     pub fn close_decimal_report(ctx: Context<CloseDecimalReport>, data_id: [u8; 16]) -> Result<()> {
-        let state = &ctx.accounts.state.load()?;
-        verify_feed_admin(&ctx.accounts.feed_admin, &state.feed_admins)?;
-
-        let feed_config = ctx.accounts.feed_config.load()?;
-
-        // if feed config workflow list is empty, then all permission accounts
-        // have also been closed as well
-        require!(
-            feed_config.workflow_metadata.is_empty(),
-            DataCacheError::FeedConfigListNotEmpty
-        );
-
-        emit!(DecimalReportClosed {
-            state: ctx.accounts.state.key(),
-            data_id: data_id
-        });
-
-        Ok(())
+        close_decimal_report::handler(ctx, data_id)
     }
 
     /// Create decimal report accounts, where report data lives (i.e answer, timestamp, etc.)
@@ -208,62 +81,7 @@ pub mod data_feeds_cache {
         ctx: Context<'_, '_, 'info, 'info, InitDecimalReports<'info>>,
         data_ids: Vec<[u8; 16]>,
     ) -> Result<()> {
-        // check feed admin here
-        let state = &ctx.accounts.state.load()?;
-        verify_feed_admin(&ctx.accounts.feed_admin, &state.feed_admins)?;
-
-        let state_key = ctx.accounts.state.key();
-
-        let data_ids_account_infos = ctx.remaining_accounts;
-
-        require_eq!(
-            data_ids.len(),
-            data_ids_account_infos.len(),
-            DataCacheError::ArrayLengthMismatch
-        );
-
-        for (i, data_id) in data_ids.iter().enumerate() {
-            let curr_report_account_info = &data_ids_account_infos[i];
-
-            let (decimal_report, bump) = Pubkey::find_program_address(
-                &[b"decimal_report", state_key.as_ref(), data_id],
-                &crate::ID,
-            );
-
-            require_keys_eq!(
-                decimal_report,
-                *curr_report_account_info.key,
-                DataCacheError::AccountMismatch
-            );
-
-            // only initialize if required
-            if curr_report_account_info.data_is_empty() {
-                let space = ANCHOR_DISCRIMINATOR + DecimalReport::INIT_SPACE;
-                let seeds: &[&[u8]] = &[b"decimal_report", state_key.as_ref(), data_id, &[bump]];
-                let payer = ctx.accounts.feed_admin.clone();
-                create_account(
-                    space,
-                    seeds,
-                    payer,
-                    curr_report_account_info.clone(),
-                    crate::ID,
-                    ctx.accounts.system_program.to_account_info(),
-                )?;
-
-                let mut dst = curr_report_account_info.try_borrow_mut_data()?;
-                dst[..ANCHOR_DISCRIMINATOR].copy_from_slice(&DecimalReport::discriminator());
-
-                emit!(DecimalReportInitialized {
-                    state: ctx.accounts.state.key(),
-                    data_id: *data_id
-                });
-            } else {
-                // verify expected format
-                DecimalReport::try_deserialize(&mut &curr_report_account_info.data.borrow()[..])?;
-            }
-        }
-
-        Ok(())
+        init_decimal_reports::handler(ctx, data_ids)
     }
 
     /// Initializes the legacy feeds config account, which stores the legacy store program
@@ -278,18 +96,7 @@ pub mod data_feeds_cache {
         ctx: Context<InitLegacyFeedsConfig>,
         data_ids: Vec<[u8; 16]>,
     ) -> Result<()> {
-        emit!(LegacyFeedsConfigInitialized {
-            state: ctx.accounts.state.key(),
-            config: ctx.accounts.legacy_feeds_config.key()
-        });
-
-        set_legacy_feeds_config(
-            ctx.accounts.legacy_feeds_config.load_init()?,
-            ctx.accounts.legacy_store.key(),
-            ctx.remaining_accounts,
-            &data_ids,
-            &vec![0_u8; data_ids.len()],
-        )
+        init_legacy_feeds_config::handler(ctx, data_ids)
     }
 
     /// Updates legacy feed config.
@@ -301,24 +108,7 @@ pub mod data_feeds_cache {
         data_ids: Vec<[u8; 16]>,
         write_disabled: Vec<bool>,
     ) -> Result<()> {
-        emit!(LegacyFeedsConfigUpdated {
-            state: ctx.accounts.state.key(),
-            config: ctx.accounts.legacy_feeds_config.key()
-        });
-
-        let write_disabled: Vec<u8> = write_disabled
-            .iter()
-            .copied() // &bool → bool
-            .map(|f| f as u8)
-            .collect();
-
-        set_legacy_feeds_config(
-            ctx.accounts.legacy_feeds_config.load_mut()?,
-            ctx.accounts.legacy_store.key(),
-            ctx.remaining_accounts,
-            &data_ids,
-            &write_disabled,
-        )
+        update_legacy_feeds_config::handler(ctx, data_ids, write_disabled)
     }
 
     /// Closes the legacy feeds config. Only to be used once all legacy feeds are
@@ -339,94 +129,7 @@ pub mod data_feeds_cache {
         descriptions: Vec<[u8; 32]>,
         workflow_metadatas: Vec<WorkflowMetadata>,
     ) -> Result<Vec<Pubkey>> {
-        validate_feed_config_inputs(&data_ids, &descriptions, &workflow_metadatas)?;
-
-        // check the remaining accounts length has sufficient feed config and permission accounts
-        let expected_len = data_ids.len() + data_ids.len() * workflow_metadatas.len();
-
-        require_eq!(
-            ctx.remaining_accounts.len(),
-            expected_len,
-            DataCacheError::InvalidAccountCount
-        );
-
-        // require ctx.remaining_accounts are in the correct order [ [...feed_config] [...permission_flags] ]
-        let feed_config_account_infos = &ctx.remaining_accounts[..data_ids.len()];
-        let permission_flag_account_infos = &ctx.remaining_accounts[data_ids.len()..];
-
-        let cache_state_key = ctx.accounts.state.key();
-
-        let mut delete_permission_accounts: Vec<Pubkey> = Vec::new();
-
-        for (i, curr_data_id) in data_ids.iter().enumerate() {
-            require!(*curr_data_id != ZERO_DATA_ID, DataCacheError::InvalidDataId);
-
-            let (curr_feed_config, _) = Pubkey::find_program_address(
-                &[b"feed_config", cache_state_key.as_ref(), curr_data_id],
-                &crate::ID,
-            );
-
-            // the feed config accounts should be in order
-            require_keys_eq!(
-                *feed_config_account_infos[i].key,
-                curr_feed_config,
-                DataCacheError::AccountMismatch
-            );
-
-            let feed_config_exists = !feed_config_account_infos[i].data_is_empty();
-
-            // sorted
-            let mut temp_candidates_deletion: Vec<Pubkey> = Vec::new();
-
-            if feed_config_exists {
-                let feed_config_loader =
-                    AccountLoader::<FeedConfig>::try_from(&feed_config_account_infos[i])?;
-
-                let feed_config = feed_config_loader.load()?;
-
-                for metadata in feed_config.workflow_metadata.iter() {
-                    // these entries are not to be deleted yet... we'll find out at the end if we need to delete them
-                    let (permission_flag, _) =
-                        permission_flag(metadata, curr_data_id, cache_state_key.as_ref());
-
-                    sorted_insert(&mut temp_candidates_deletion, permission_flag)
-                }
-            }
-
-            for (j, metadata) in workflow_metadatas.iter().enumerate() {
-                let (curr_permission_flag, _) =
-                    permission_flag(metadata, curr_data_id, cache_state_key.as_ref());
-
-                // ex: data_ids: [1, 2]
-                // workflow metadatas [5, 6, 7]
-                // ctx remaining accounts:
-                // [1-feed-config]  |- feed_config_accounts
-                // [2-feed-config]  |
-                // [flag-1-5] [flag-1-6] [flag-1-7]  |- permission_flag_accounts
-                // [flag-2-5] [flag-2-6] [flag-2-7]  |
-
-                let permission_flag_account_info =
-                    &permission_flag_account_infos[i * workflow_metadatas.len() + j];
-
-                // check that it is in the remaining accounts
-                require_keys_eq!(
-                    curr_permission_flag,
-                    *permission_flag_account_info.key,
-                    DataCacheError::AccountMismatch
-                );
-
-                // permission flag are removed from deletion set because it's still in use
-                if let Ok(index) = temp_candidates_deletion.binary_search(&curr_permission_flag) {
-                    temp_candidates_deletion.remove(index);
-                }
-            }
-
-            // add items
-            delete_permission_accounts.append(&mut temp_candidates_deletion);
-        }
-
-        // order has to be exactly the same
-        Ok(delete_permission_accounts)
+        preview_decimal_feed_configs::handler(ctx, data_ids, descriptions, workflow_metadatas)
     }
 
     /// Given N data ids and M workflows, configures N feed config accounts and N*M write permission accounts.
@@ -442,210 +145,7 @@ pub mod data_feeds_cache {
         descriptions: Vec<[u8; 32]>,
         workflow_metadatas: Vec<WorkflowMetadata>,
     ) -> Result<()> {
-        // check feed admin here
-        let state = &mut ctx.accounts.state.load()?;
-        verify_feed_admin(&ctx.accounts.feed_admin, &state.feed_admins)?;
-
-        validate_feed_config_inputs(&data_ids, &descriptions, &workflow_metadatas)?;
-
-        // check the remaining accounts length has sufficient feed config and permission accounts
-        let minimum_len = data_ids.len() + data_ids.len() * workflow_metadatas.len();
-
-        // you have an unknown of defunct permission accounts as well, so as long as the amount is >= we're good
-        require_gte!(
-            ctx.remaining_accounts.len(),
-            minimum_len,
-            DataCacheError::InvalidAccountCount
-        );
-
-        // require ctx.remaining_accounts are in the correct order [ [...feed_config] [...permission_flags] ]
-        let feed_config_account_infos = &ctx.remaining_accounts[..data_ids.len()];
-        let index = data_ids.len() + data_ids.len() * workflow_metadatas.len();
-        let permission_flag_account_infos = &ctx.remaining_accounts[data_ids.len()..index];
-        let delete_permission_account_infos = &ctx.remaining_accounts[index..];
-
-        let cache_state_key = ctx.accounts.state.key();
-
-        let mut delete_permission_accounts: Vec<Pubkey> = Vec::new();
-
-        for (i, curr_data_id) in data_ids.iter().enumerate() {
-            require!(*curr_data_id != ZERO_DATA_ID, DataCacheError::InvalidDataId);
-
-            // derive the PDA
-            // get the existing config feed, see if it's empty or not
-            let (curr_feed_config, feed_config_bump) = Pubkey::find_program_address(
-                &[b"feed_config", cache_state_key.as_ref(), curr_data_id],
-                &crate::ID,
-            );
-
-            // the feed config accounts should be in order
-            require_keys_eq!(
-                *feed_config_account_infos[i].key,
-                curr_feed_config,
-                DataCacheError::AccountMismatch
-            );
-
-            let feed_config_exists = !feed_config_account_infos[i].data_is_empty();
-
-            let feed_config_loader = if feed_config_exists {
-                AccountLoader::<FeedConfig>::try_from(&feed_config_account_infos[i])?
-            } else {
-                let space = ANCHOR_DISCRIMINATOR + FeedConfig::INIT_SPACE;
-                // initialize it
-                let seeds: &[&[u8]] = &[
-                    b"feed_config",
-                    cache_state_key.as_ref(),
-                    curr_data_id,
-                    &[feed_config_bump],
-                ];
-
-                create_account(
-                    space,
-                    seeds,
-                    ctx.accounts.feed_admin.clone(),
-                    feed_config_account_infos[i].clone(),
-                    crate::ID,
-                    ctx.accounts.system_program.to_account_info(),
-                )?;
-
-                // avoid double borrow to write discriminator
-                {
-                    let mut dst = feed_config_account_infos[i].try_borrow_mut_data()?;
-                    dst[..ANCHOR_DISCRIMINATOR].copy_from_slice(&FeedConfig::discriminator());
-                }
-
-                AccountLoader::<FeedConfig>::try_from(&feed_config_account_infos[i])?
-            };
-
-            // load_mut instead of load_init because we write the discriminator above
-            let mut feed_config = feed_config_loader.load_mut()?;
-
-            // sorted
-            let mut temp_candidates_deletion: Vec<Pubkey> = Vec::new();
-
-            // so these are the permission accounts you need to delete later
-            if feed_config_exists {
-                // go over current workflows
-                for metadata in feed_config.workflow_metadata.iter() {
-                    // these entries are not to be deleted yet... we'll find out at the end if we need to delete them
-                    let (permission_flag, _) =
-                        permission_flag(metadata, curr_data_id, cache_state_key.as_ref());
-
-                    sorted_insert(&mut temp_candidates_deletion, permission_flag);
-                }
-            }
-
-            // let mut new_workflow_metadata = Vec::default();
-
-            feed_config.workflow_metadata.clear();
-
-            // go over new workflows to be added
-            // inner loop iterates over the workflow metadata
-            for (j, metadata) in workflow_metadatas.iter().enumerate() {
-                let report_hash = create_report_hash(
-                    curr_data_id,
-                    &metadata.allowed_sender,
-                    &metadata.allowed_workflow_owner,
-                    &metadata.allowed_workflow_name,
-                );
-                let (curr_permission_flag, bump) = Pubkey::find_program_address(
-                    &[
-                        b"permission_flag",
-                        ctx.accounts.state.key().as_ref(),
-                        &report_hash,
-                    ],
-                    &crate::ID,
-                );
-
-                // ex: data_ids: [1, 2]
-                // workflow metadatas [5, 6, 7]
-                // ctx remaining accounts:
-                // [1-feed-config]  |- feed_config_accounts
-                // [2-feed-config]  |
-                // [flag-1-5] [flag-1-6] [flag-1-7]  |- permission_flag_accounts
-                // [flag-2-5] [flag-2-6] [flag-2-7]  |
-
-                let permission_flag_account_info =
-                    &permission_flag_account_infos[i * workflow_metadatas.len() + j];
-
-                // check that it is in the remaining accounts
-                require_keys_eq!(
-                    curr_permission_flag,
-                    *permission_flag_account_info.key,
-                    DataCacheError::AccountMismatch
-                );
-
-                // create permission_flag if needed
-                if permission_flag_account_info.data_is_empty() {
-                    let seeds: &[&[u8]] = &[
-                        b"permission_flag",
-                        cache_state_key.as_ref(),
-                        &report_hash,
-                        &[bump],
-                    ];
-
-                    create_account(
-                        ANCHOR_DISCRIMINATOR,
-                        seeds,
-                        ctx.accounts.feed_admin.clone(),
-                        permission_flag_account_info.clone(),
-                        crate::ID,
-                        ctx.accounts.system_program.to_account_info(),
-                    )?;
-
-                    let mut dst = permission_flag_account_info.try_borrow_mut_data()?;
-                    dst[..ANCHOR_DISCRIMINATOR]
-                        .copy_from_slice(&WritePermissionFlag::discriminator());
-                }
-
-                // ensure the flag has expected schema
-                WritePermissionFlag::try_deserialize(
-                    &mut &permission_flag_account_info.data.borrow()[..],
-                )?;
-
-                // permission flag are removed from deletion set because it's still in use
-                if let Ok(index) = temp_candidates_deletion.binary_search(&curr_permission_flag) {
-                    temp_candidates_deletion.remove(index);
-                }
-
-                feed_config.workflow_metadata.push(*metadata);
-            }
-
-            feed_config.description = descriptions[i];
-
-            emit!(DecimalFeedConfigSet {
-                state: ctx.accounts.state.key(),
-                data_id: *curr_data_id,
-                decimals: get_decimals(curr_data_id),
-                description: descriptions[i],
-                workflow_metadatas: workflow_metadatas.clone(),
-            });
-
-            delete_permission_accounts.append(&mut temp_candidates_deletion)
-        }
-
-        require_eq!(
-            delete_permission_accounts.len(),
-            delete_permission_account_infos.len(),
-            DataCacheError::ArrayLengthMismatch
-        );
-
-        for (i, permission_account) in delete_permission_accounts.iter().enumerate() {
-            let curr_permission_account_info = &delete_permission_account_infos[i];
-
-            require_keys_eq!(
-                *permission_account,
-                *curr_permission_account_info.key,
-                DataCacheError::AccountMismatch
-            );
-
-            close_account(
-                curr_permission_account_info.clone(),
-                ctx.accounts.feed_admin.to_account_info(),
-            )?;
-        }
-
-        Ok(())
+        set_decimal_feed_configs::handler(ctx, data_ids, descriptions, workflow_metadatas)
     }
 
     /// Receives the a forwarder report that contains a list of [`ReceivedDecimalReport`]
@@ -660,283 +160,7 @@ pub mod data_feeds_cache {
         metadata: Vec<u8>,
         report: Vec<u8>,
     ) -> Result<()> {
-        let legacy_feeds_config = if let Some(loader) = &ctx.accounts.legacy_feeds_config {
-            let loader = loader.load()?;
-
-            // if included, check that the legacy store passed in via account context is the same as the one in the config
-            if let Some(legacy_store) = &ctx.accounts.legacy_store {
-                require_keys_eq!(
-                    *legacy_store.key,
-                    loader.legacy_store,
-                    DataCacheError::AccountMismatch
-                );
-            };
-
-            Some(loader)
-        } else {
-            None
-        };
-
-        // first assume we don't have legacy_store or legacy_feed_config
-
-        let (workflow_name, workflow_owner) = get_workflow_metadata(&metadata)?;
-
-        let received_decimal_reports = Vec::<ReceivedDecimalReport>::try_from_slice(&report)
-            .map_err(|_| DataCacheError::MalformedReport)?;
-
-        let len = received_decimal_reports.len();
-
-        let report_account_infos = &ctx.remaining_accounts[..len];
-        let permission_flag_account_infos = &ctx.remaining_accounts[len..2 * len];
-        let legacy_feed_account_infos = &ctx.remaining_accounts[2 * len..];
-
-        // sorted by key
-        let legacy_accounts_sorted = legacy_feed_account_infos
-            .windows(2)
-            .all(|w| w[0].key.lt(w[1].key));
-
-        require!(
-            legacy_accounts_sorted,
-            DataCacheError::AddressesMustStrictlyIncrease
-        );
-
-        require_eq!(
-            report_account_infos.len(),
-            received_decimal_reports.len(),
-            DataCacheError::ArrayLengthMismatch
-        );
-
-        require_eq!(
-            permission_flag_account_infos.len(),
-            received_decimal_reports.len(),
-            DataCacheError::ArrayLengthMismatch
-        );
-
-        let mut candidate_legacy_writes: Vec<(&LegacyFeedEntry, &ReceivedDecimalReport)> =
-            Vec::new();
-
-        for (i, received_decimal_report) in received_decimal_reports.iter().enumerate() {
-            // 1. check that sender has permission to write
-            let report_hash = create_report_hash(
-                &received_decimal_report.data_id,
-                ctx.accounts.forwarder_authority.key,
-                workflow_owner,
-                workflow_name,
-            );
-
-            let (curr_permission_flag, _) = Pubkey::find_program_address(
-                &[
-                    b"permission_flag",
-                    ctx.accounts.cache_state.key().as_ref(),
-                    &report_hash,
-                ],
-                &crate::ID,
-            );
-
-            require_keys_eq!(
-                curr_permission_flag,
-                *permission_flag_account_infos[i].key,
-                DataCacheError::AccountMismatch
-            );
-
-            // verifies the permission account exists
-            if WritePermissionFlag::try_deserialize(
-                &mut &permission_flag_account_infos[i].data.borrow()[..],
-            )
-            .is_err()
-            {
-                emit!(InvalidUpdatePermission {
-                    state: ctx.accounts.cache_state.key(),
-                    data_id: received_decimal_report.data_id,
-                    sender: ctx.accounts.forwarder_authority.key(),
-                    workflow_owner: workflow_owner
-                        .try_into()
-                        .map_err(|_| DataCacheError::InvalidLength)?,
-                    workflow_name: workflow_name
-                        .try_into()
-                        .map_err(|_| DataCacheError::InvalidLength)?,
-                });
-
-                continue;
-            }
-
-            // 2. check report account is valid
-            let (curr_report, _) = Pubkey::find_program_address(
-                &[
-                    b"decimal_report",
-                    ctx.accounts.cache_state.key().as_ref(),
-                    &received_decimal_report.data_id,
-                ],
-                &crate::ID,
-            );
-
-            require_keys_eq!(
-                curr_report,
-                *report_account_infos[i].key,
-                DataCacheError::AccountMismatch
-            );
-
-            // update report
-
-            let latest_report =
-                DecimalReport::try_deserialize(&mut &report_account_infos[i].data.borrow()[..])?;
-
-            // don't update if the received report is stale
-            if received_decimal_report.timestamp <= latest_report.timestamp {
-                emit!(StaleDecimalReport {
-                    state: ctx.accounts.cache_state.key(),
-                    data_id: received_decimal_report.data_id,
-                    received_timestamp: received_decimal_report.timestamp,
-                    latest_timestamp: latest_report.timestamp
-                });
-
-                continue;
-            }
-
-            let mut dst = report_account_infos[i].try_borrow_mut_data()?;
-
-            let updated_report = DecimalReport {
-                answer: received_decimal_report.answer,
-                timestamp: received_decimal_report.timestamp,
-            };
-
-            updated_report.serialize(&mut &mut dst[ANCHOR_DISCRIMINATOR..])?;
-
-            emit!(DecimalReportUpdated {
-                state: ctx.accounts.cache_state.key(),
-                answer: received_decimal_report.answer,
-                timestamp: received_decimal_report.timestamp,
-                data_id: received_decimal_report.data_id
-            });
-
-            // 3. check if the report is also associated with a legacy feed
-            // a. search config by data_id to get the account key
-            // b. search passed in legacy_feed_account_infos by key
-
-            if let Some(config) = &legacy_feeds_config {
-                // a given legacy feed will only write under conditions
-                // I. legacy feed config is provided
-                // II. data id is associated with a legacy feed in the config
-                // III. the legacy store is provided
-                // IV. legacy writer is provided
-                // V. writes are not disabled for that legacy feed
-                // VI. the legacy feed is provided in account context
-
-                // condition I and II: if the data id is associated with a legacy feed
-                if let Some(entry) = config
-                    .id_to_feed
-                    .binary_search_by(|e| e.data_id.cmp(&received_decimal_report.data_id))
-                    .ok()
-                    .and_then(|index| config.id_to_feed.get(index))
-                {
-                    candidate_legacy_writes.push((entry, received_decimal_report));
-                }
-            }
-        }
-
-        // or add generic dfc event out here
-
-        // seperate out write disabled entries
-        let (write_disabled_entries, write_enabled_entries): (
-            Vec<(&LegacyFeedEntry, &ReceivedDecimalReport)>,
-            Vec<(&LegacyFeedEntry, &ReceivedDecimalReport)>,
-        ) = candidate_legacy_writes
-            .iter()
-            .partition(|e| e.0.write_disabled != 0);
-
-        let mut write_occurred = false;
-
-        // condition III & condition IV
-        if let (Some(legacy_store), Some(legacy_writer)) =
-            (&ctx.accounts.legacy_store, &ctx.accounts.legacy_writer)
-        {
-            // use legacy_store and legacy_writer here
-
-            let mut ordered_legacy_feed_account_infos: Vec<&AccountInfo> = Vec::new();
-
-            // condition V
-            for entry in write_enabled_entries.iter() {
-                // condition VI: error if legacy feed account not supplied in account context
-                let account = legacy_feed_account_infos
-                    .binary_search_by(|a| a.key.cmp(&entry.0.legacy_feed))
-                    .map(|i| &legacy_feed_account_infos[i])
-                    .map_err(|_| DataCacheError::MissingLegacyFeedAccount)?;
-
-                ordered_legacy_feed_account_infos.push(account);
-            }
-
-            // write to store program
-            if !write_enabled_entries.is_empty() {
-                let metas: Vec<AccountMeta> = std::iter::once(AccountMeta {
-                    pubkey: legacy_writer.key(),
-                    is_signer: true,
-                    is_writable: false,
-                })
-                .chain(
-                    ordered_legacy_feed_account_infos
-                        .iter()
-                        .map(|acc| AccountMeta {
-                            pubkey: *acc.key,
-                            is_signer: false,
-                            is_writable: true,
-                        }),
-                )
-                .collect();
-
-                let account_infos: Vec<AccountInfo<'info>> =
-                    std::iter::once(legacy_writer.to_account_info())
-                        .chain(
-                            ordered_legacy_feed_account_infos
-                                .iter()
-                                .map(|val| val.to_account_info()),
-                        )
-                        .collect();
-
-                // payload begins with the Anchor discriminator
-                let mut payload = SUBMIT_DISCRIMINATOR.to_vec();
-
-                let transmissions: Vec<CacheTransmission> = write_enabled_entries
-                    .iter()
-                    .map(|e| CacheTransmission {
-                        timestamp: e.1.timestamp,
-                        answer: e.1.answer,
-                    })
-                    .collect();
-
-                payload.extend(transmissions.try_to_vec()?);
-
-                let cache_state_key = ctx.accounts.cache_state.key();
-
-                let ix = Instruction::new_with_bytes(legacy_store.key(), &payload, metas);
-                let signer_seeds = &[
-                    b"legacy_writer",
-                    cache_state_key.as_ref(),
-                    &[ctx.accounts.cache_state.load()?.legacy_writer_bump],
-                ];
-
-                invoke_signed(&ix, &account_infos, &[signer_seeds])
-                    .map_err(|_| DataCacheError::FailedLegacyWrite)?;
-
-                write_occurred = true;
-            }
-        }
-
-        // emit legacy event only if there were candidates identified by the feed config
-        if !candidate_legacy_writes.is_empty() {
-            let (feeds_skipped, feeds_written) = if write_occurred {
-                (write_disabled_entries, write_enabled_entries)
-            } else {
-                (candidate_legacy_writes, vec![])
-            };
-
-            emit!(LegacyFeedsReported {
-                state: ctx.accounts.cache_state.key(),
-                feeds_skipped: feeds_skipped.iter().map(|e| e.0.data_id).collect(),
-                feeds_written: feeds_written.iter().map(|e| e.0.data_id).collect(),
-            });
-        }
-
-        Ok(())
+        on_report::handler(ctx, metadata, report)
     }
 
     /// Returns a feed's workflow metadata. Chunks return
@@ -952,32 +176,7 @@ pub mod data_feeds_cache {
         start_index: u8,
         max_count: u8,
     ) -> Result<Vec<WorkflowMetadata>> {
-        let feed_config = ctx.accounts.feed_config.load()?;
-
-        require!(
-            !feed_config.workflow_metadata.is_empty(),
-            DataCacheError::FeedNotConfigured
-        );
-
-        let len = feed_config.workflow_metadata.len();
-
-        let start_index: usize = start_index.into();
-        let max_count: usize = max_count.into();
-
-        if start_index >= len {
-            return Ok(Vec::new());
-        }
-
-        // max count 0 means take start_index and everything after it
-
-        let mut end_index = start_index + max_count;
-        end_index = if end_index > len || max_count == 0 {
-            len
-        } else {
-            end_index
-        };
-
-        Ok(feed_config.workflow_metadata[start_index..end_index].to_vec())
+        query_feed_metadata::handler(ctx, _data_id, start_index, max_count)
     }
 
     /// The data ids passed in must match the ordering of the decimal report accounts
@@ -986,253 +185,6 @@ pub mod data_feeds_cache {
         ctx: Context<'_, '_, 'info, 'info, QueryValues<'info>>,
         data_ids: Vec<[u8; 16]>,
     ) -> Result<Vec<DecimalReport>> {
-        require_eq!(
-            data_ids.len(),
-            ctx.remaining_accounts.len(),
-            DataCacheError::ArrayLengthMismatch
-        );
-
-        let mut reports = Vec::new();
-
-        for (i, data_id) in data_ids.iter().enumerate() {
-            let (decimal_report, _) = Pubkey::find_program_address(
-                &[
-                    b"decimal_report",
-                    ctx.accounts.cache_state.key().as_ref(),
-                    data_id,
-                ],
-                &crate::ID,
-            );
-
-            let report_account_info = &ctx.remaining_accounts[i];
-
-            require_keys_eq!(
-                decimal_report,
-                *report_account_info.key,
-                DataCacheError::AccountMismatch
-            );
-
-            let r = Account::<DecimalReport>::try_from(report_account_info)?;
-
-            reports.push(DecimalReport {
-                timestamp: r.timestamp,
-                answer: r.answer,
-            });
-        }
-
-        Ok(reports)
+        query_values::handler(ctx, data_ids)
     }
-}
-
-// already checked it's empty
-fn create_account<'info>(
-    space: usize,
-    seeds: &[&[u8]],
-    payer: Signer<'info>,
-    to_account: AccountInfo<'info>,
-    program_id: Pubkey,
-    system_program: AccountInfo<'info>,
-) -> Result<()> {
-    let rent = Rent::get()?.minimum_balance(space);
-
-    let current_lamports = to_account.lamports();
-    if current_lamports == 0 {
-        invoke_signed(
-            &system_instruction::create_account(
-                payer.key,
-                to_account.key,
-                rent,
-                space as u64,
-                &program_id,
-            ),
-            &[
-                payer.to_account_info(),
-                to_account.clone(),
-                system_program.clone(),
-            ],
-            &[seeds],
-        )?;
-    } else {
-        // do extra stuff
-        let required_lamports = rent.saturating_sub(current_lamports);
-        // transfer remaining lamports
-        if required_lamports > 0 {
-            let cpi_accounts = anchor_lang::system_program::Transfer {
-                from: payer.to_account_info(),
-                to: to_account.clone(),
-            };
-            let cpi_context =
-                anchor_lang::context::CpiContext::new(system_program.clone(), cpi_accounts);
-            anchor_lang::system_program::transfer(cpi_context, required_lamports)?;
-        }
-        // allocate space
-        let cpi_accounts = anchor_lang::system_program::Allocate {
-            account_to_allocate: to_account.clone(),
-        };
-        let cpi_context =
-            anchor_lang::context::CpiContext::new(system_program.clone(), cpi_accounts);
-        anchor_lang::system_program::allocate(cpi_context.with_signer(&[seeds]), space as u64)?;
-
-        // Assign ownership to program
-        let cpi_accounts = anchor_lang::system_program::Assign {
-            account_to_assign: to_account.clone(),
-        };
-        let cpi_context =
-            anchor_lang::context::CpiContext::new(system_program.clone(), cpi_accounts);
-        anchor_lang::system_program::assign(cpi_context.with_signer(&[seeds]), &program_id)?;
-    }
-    Ok(())
-}
-
-fn verify_feed_admin(admin: &Signer, admin_list: &AccountList) -> Result<()> {
-    let is_admin = admin_list.binary_search(admin.key).is_ok();
-    require!(is_admin, AuthError::Unauthorized);
-
-    Ok(())
-}
-
-fn create_report_hash(data_id: &[u8], sender: &Pubkey, owner: &[u8], name: &[u8]) -> [u8; 32] {
-    hash::hash(&[data_id, &sender.to_bytes(), owner, name].concat()).to_bytes()
-}
-
-fn get_decimals(data_id: &[u8; 16]) -> u8 {
-    let report_type = data_id[7];
-
-    if (0x20..=0x60).contains(&report_type) {
-        report_type - 32
-    } else {
-        0
-    }
-}
-
-fn close_account(account: AccountInfo, destination: AccountInfo) -> Result<()> {
-    **destination.lamports.borrow_mut() = destination
-        .lamports()
-        .checked_add(account.lamports())
-        .unwrap();
-    **account.lamports.borrow_mut() = 0;
-
-    let mut data = account.try_borrow_mut_data()?;
-    for byte in data.deref_mut().iter_mut() {
-        *byte = 0;
-    }
-
-    Ok(())
-}
-
-// workflow_cid           offset  0, size 32
-// workflow_name          offset  32, size 10
-// workflow_owner         offset  42, size 20
-// report_id              offset  62, size  2
-fn get_workflow_metadata(metadata: &[u8]) -> Result<(&[u8], &[u8])> {
-    let workflow_name = metadata.get(32..42).ok_or(DataCacheError::OutOfBounds)?;
-    let workflow_owner = metadata.get(42..62).ok_or(DataCacheError::OutOfBounds)?;
-
-    Ok((workflow_name, workflow_owner))
-}
-
-fn set_legacy_feeds_config(
-    mut legacy_feeds_config: RefMut<LegacyFeedsConfig>,
-    legacy_store: Pubkey,
-    legacy_feeds: &[AccountInfo],
-    data_ids: &[[u8; 16]],
-    write_disabled: &[u8],
-) -> Result<()> {
-    require!(
-        data_ids.len() == legacy_feeds.len() && data_ids.len() == write_disabled.len(),
-        DataCacheError::ArrayLengthMismatch
-    );
-
-    // reset the array
-    legacy_feeds_config.id_to_feed.clear();
-
-    legacy_feeds_config.legacy_store = legacy_store;
-
-    let mut prev_data_id = [0_u8; 16];
-    for (i, data_id) in data_ids.iter().enumerate() {
-        require!(
-            prev_data_id < *data_id,
-            DataCacheError::IdsMustStrictlyIncrease
-        );
-
-        legacy_feeds_config.id_to_feed.push(LegacyFeedEntry {
-            data_id: *data_id,
-            legacy_feed: legacy_feeds[i].key(),
-            write_disabled: write_disabled[i],
-        });
-
-        prev_data_id = *data_id;
-    }
-
-    Ok(())
-}
-
-fn sorted_insert<T: Ord>(vec: &mut Vec<T>, value: T) {
-    match vec.binary_search(&value) {
-        Ok(pos) | Err(pos) => vec.insert(pos, value),
-    }
-}
-
-fn validate_feed_config_inputs(
-    data_ids: &Vec<[u8; 16]>,
-    descriptions: &Vec<[u8; 32]>,
-    workflow_metadatas: &Vec<WorkflowMetadata>,
-) -> Result<()> {
-    require_gte!(
-        MAX_WORKFLOW_METADATAS,
-        workflow_metadatas.len(),
-        DataCacheError::MaxWorkflowsExceeded
-    );
-
-    require!(!descriptions.is_empty(), DataCacheError::EmptyConfig);
-
-    require_eq!(
-        data_ids.len(),
-        descriptions.len(),
-        DataCacheError::ArrayLengthMismatch
-    );
-
-    for d in descriptions.iter() {
-        if workflow_metadatas.is_empty() {
-            require!(*d == [0; 32], DataCacheError::EmptyDescriptionEnforced);
-        } else {
-            require!(*d != [0; 32], DataCacheError::InvalidDescrption);
-        }
-    }
-
-    for metadata in workflow_metadatas.iter() {
-        require_keys_neq!(
-            metadata.allowed_sender,
-            Pubkey::default(),
-            DataCacheError::InvalidAddress
-        );
-        require!(
-            !metadata.allowed_workflow_name.is_empty(),
-            DataCacheError::InvalidWorkflowName
-        );
-        require!(
-            metadata.allowed_workflow_owner != ZERO_ADDRESS,
-            DataCacheError::InvalidAddress
-        );
-    }
-
-    Ok(())
-}
-
-fn permission_flag(
-    metadata: &WorkflowMetadata,
-    data_id: &[u8; 16],
-    cache_state: &[u8],
-) -> (Pubkey, u8) {
-    let derived_report_hash = create_report_hash(
-        data_id,
-        &metadata.allowed_sender,
-        &metadata.allowed_workflow_owner,
-        &metadata.allowed_workflow_name,
-    );
-
-    Pubkey::find_program_address(
-        &[b"permission_flag", cache_state, &derived_report_hash],
-        &crate::ID,
-    )
 }
